Speed Update

+++ config/repo-init.ts
@@ -7,16 +7,14 @@ const rl = readline.createInterface({
     output: process.stdout
 });
 
-fs.writeFileSync('messages.json', '[]', 'utf8');
-fs.writeFileSync('webhooks.json', '[]', 'utf8');
-fs.writeFileSync('userAuths.json', '{}', 'utf8');
+if (!fs.existsSync('messages.json')) fs.writeFileSync('messages.json', '[]', 'utf8');
+if (!fs.existsSync('webhooks.json')) fs.writeFileSync('webhooks.json', '[]', 'utf8');
+if (!fs.existsSync('userAuths.json')) fs.writeFileSync('userAuths.json', '{}', 'utf8');
 
 console.log("A gmail account with less secure apps access turned on is required to run the chat.\nPlease enter login details for that account:");
 rl.question('Enter email address: ', email => {
     rl.question('Enter email password: ', password => {
         fs.writeFileSync('.env', `EMAIL_PASS=${password}\nEMAIL=${email}`);
-        return;
+        process.exit();
     });
 });
-
-exec('npm run build');
